registro, validacion de usuarios duplicados
