+++ db.json
@@ -26,6 +26,69 @@
       "telefono": "71816948",
       "role": "0",
       "password": "sIcARIo2004$hi$red#"
+    },
+    {
+      "id": "",
+      "nombre": "jeff",
+      "apellido": "",
+      "correo": "",
+      "password": {}
+    },
+    {
+      "id": "456789",
+      "nombre": "jeff",
+      "apellido": "medina",
+      "correo": "jef@gmail.com",
+      "password": {}
+    },
+    {
+      "id": "456789",
+      "nombre": "jeff",
+      "apellido": "medina",
+      "correo": "jef@gmail.com",
+      "password": {}
+    },
+    {
+      "id": "12345",
+      "nombre": "jeff",
+      "apellido": "oco",
+      "correo": "jef@gmail.com",
+      "password": {}
+    },
+    {
+      "id": "2324",
+      "nombre": "gee",
+      "apellido": "jjj",
+      "correo": "def@gmail.com",
+      "password": "355"
+    },
+    {
+      "id": "604700161",
+      "nombre": "jeffry",
+      "apellido": "Medina",
+      "correo": "jef@gmail.com",
+      "password": "12345"
+    },
+    {
+      "id": "604700161",
+      "nombre": "jeffry",
+      "apellido": "Medina",
+      "correo": "jef@gmail.com",
+      "password": "12345"
+    },
+    {
+      "id": "1234",
+      "nombre": "andres",
+      "apellido": "medina",
+      "correo": "jef@gmail.com",
+      "password": "1"
+    },
+    {
+      "id": "13",
+      "nombre": "jeffryk",
+      "apellido": "medina",
+      "correo": "je22223hf@gmail.com",
+      "password": "12"
     }
   ],
   "peticiones": [
@@ -77,6 +140,15 @@
       "serie": "99999",
       "marca": "msi",
       "fechaSalida": "2024-08-24",
+      "fechaRegreso": "2024-08-25",
+      "estado": "0",
+      "id-user": "60888088"
+    },
+    {
+      "id": "9925",
+      "serie": "nrt980",
+      "marca": "Pinocho",
+      "fechaSalida": "2024-08-23",
       "fechaRegreso": "2024-08-25",
       "estado": "0",
       "id-user": "60888088"
