+++ package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "npx parcel public/index.html --port 5000",
+    "start": "npx parcel public/index.html --port 8080",
     "server": "json-server --watch db.json --port 3000",
     "api": ""
   },
