+++ public/css/styles.css
@@ -348,3 +348,11 @@ img{
     }
     
 }
+.Registro{
+    margin-left: 400px;
+    margin-right: 400px;
+    border: 1px solid;
+}
+label{
+
+}
