+++ public/register.html
@@ -32,31 +32,31 @@
         <div class="wrapper  d-flex align-items-center justify-content-center h-100">
             <div class="card login-form register-container">
                 <div class="card-body">
-                    <h5 class="card-title text-center">Register Form</h5>
+                    <h5 class="card-title text-center">Formulario de Registro</h5>
                     <form id="registerForm">
                         <div class="form-group mb-3">
-                            <label for="" class="form-label">First Name</label>
-                            <input type="text" class="form-control" id="firstName" name="firstName" placeholder="John">
+                            <label for="" class="form-label">Nombre</label>
+                            <input type="text" class="form-control" id="Nombre" name="Nombre" placeholder="John">
                         </div>
                         <div class="form-group mb-3">
-                            <label for="" class="form-label">Last Name</label>
-                            <input type="text" class="form-control" id="lastName"  name="lastName" placeholder="Smith">
+                            <label for="" class="form-label">Apellidos</label>
+                            <input type="text" class="form-control" id="Apellidos"  name="Apellidos" placeholder="Smith">
                         </div>
                         <div class="form-group mb-3">
-                            <label for="" class="form-label">Phone Number</label>
-                            <input type="text" class="form-control" id="phoneNumber"  name="phoneno" placeholder="888 104 365">
+                            <label for="" class="form-label">Numero de Cedula:</label>
+                            <input type="text" class="form-control" id="NumeroCedula"  name="phoneno" placeholder="888 104 365">
                         </div>
                         <div class="form-group mb-3">
-                            <label for="" class="form-label">Email address</label>
-                            <input type="email" class="form-control" id="email" aria-describedby="emailHelp" name="email" placeholder="example@email.com">
+                            <label for="" class="form-label">Correo Electronico</label>
+                            <input type="email" class="form-control" id="CorreoElectronico" aria-describedby="emailHelp" name="email" placeholder="example@email.com">
                         </div>
                         <div class="form-group mb-3">
-                            <label for="" class="form-label">Password</label>
-                            <input type="password" class="form-control" id="password" name="password" placeholder="Secret Password">
+                            <label for="" class="form-label">Contraseña</label>
+                            <input type="password" class="form-control" id="Contraseña" name="Contraseña" placeholder="12345">
                         </div>
-                        <button type="submit" class="btn btn-primary w-100" name="create">Sign up</button>
+                        <button type="submit" class="btn btn-primary w-100" id="Registrar" name="create">Registrar</button>
                         <div class="sign-up mt-4">
-                            Already have an account? <a href="login.html">Login</a>
+                            Y tienes una Cuenta? <a href="login.html">Inicio Sesion</a>
                         </div>
                     </form>
                 </div>
@@ -73,6 +73,7 @@
 
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-<script type="module" src="/services/login/login.js"></script>
+
+<script type="module" src="./js/register.js"></script>
 </body>
 </html>
