+++ public/solicitud.html
@@ -58,5 +58,6 @@
     </div>
 
     <script type="module" src="./js/solicitud.js"></script>
+    
 </body>
 </html>
